requirements.txt
pandas
scikit-learn
nba_api
matplotlib
tabulate
Flask
pyyaml